+++ packages/at_secondary_server/lib/src/verb/handler/keys_verb_handler.dart
@@ -16,147 +16,171 @@ class KeysVerbHandler extends AbstractVerbHandler {
 
   KeysVerbHandler(SecondaryKeyStore keyStore) : super(keyStore);
 
-  // Method to verify whether command is accepted or not
-  // Input: command
   @override
   bool accept(String command) => command.startsWith('keys:');
 
-  // Method to return Instance of verb belongs to this VerbHandler
   @override
   Verb getVerb() {
     return keys;
   }
 
-  // Method which will process update Verb
-  // This will process given verb and write response to response object
-  // Input : Response, verbParams, AtConnection
   @override
   Future<void> processVerb(
-      Response response,
-      HashMap<String, String?> verbParams,
-      InboundConnection atConnection) async {
+    Response response,
+    HashMap<String, String?> verbParams,
+    InboundConnection atConnection,
+  ) async {
     final keyVisibility = verbParams[visibility];
     final atSign = AtSecondaryServerImpl.getInstance().currentAtSign;
     bool hasManageAccess = false;
     var connectionMetadata =
         atConnection.getMetaData() as InboundConnectionMetadata;
     final enrollIdFromMetadata = connectionMetadata.enrollmentId;
+    if (enrollIdFromMetadata == null) {
+      throw AtEnrollmentException(
+          'Keys verb cannot be accessed without an enrollmentId');
+    }
     logger.finer('enrollIdFromMetadata:$enrollIdFromMetadata');
     final key =
         '$enrollIdFromMetadata.$newEnrollmentKeyPattern.$enrollManageNamespace';
-    var enrollData;
-    try {
-      enrollData = await keyStore.get('$key$atSign');
-    } on KeyNotFoundException {
-      logger.warning('enrollment key not found in keystore $key');
-      throw AtEnrollmentException(
-          'Enrollment Id $enrollIdFromMetadata not found in keystore');
-    }
+
+    var enrollData = await _getEnrollData(key, atSign);
     if (enrollData != null) {
       final enrollDataStoreValue =
-          EnrollDataStoreValue.fromJson(jsonDecode(enrollData.data));
-      if (enrollDataStoreValue.approval!.state != 'approved') {
+          EnrollDataStoreValue.fromJson(jsonDecode(enrollData.data!));
+      if (enrollDataStoreValue.approval?.state != 'approved') {
         throw AtEnrollmentException(
-            'Enrollment Id $enrollmentId is not approved. current state :${enrollDataStoreValue.approval!.state}');
+            'Enrollment Id $enrollIdFromMetadata is not approved. current state: ${enrollDataStoreValue.approval?.state}');
       }
       hasManageAccess =
-          enrollDataStoreValue.namespaces.containsKey(enrollManageNamespace) &&
-              enrollDataStoreValue.namespaces[enrollManageNamespace] == 'rw';
+          enrollDataStoreValue.namespaces[enrollManageNamespace] == 'rw';
     }
+
     final value = verbParams[keyValue];
-    final valueJson = {};
-    valueJson['value'] = value;
-    valueJson['keyType'] = verbParams[keyType];
-    valueJson[enrollmentId] = enrollIdFromMetadata;
+    final valueJson = {
+      'value': value,
+      'keyType': verbParams[keyType],
+      enrollmentId: enrollIdFromMetadata
+    };
     final operation = verbParams[AT_OPERATION];
+
     switch (operation) {
       case 'put':
-        try {
-          logger.finer('value:$valueJson');
-          // update the key in data store
-          final atData = AtData();
-          atData.data = jsonEncode(valueJson);
-          dynamic result;
-          var keyName = _getKeyName(verbParams, atSign, keyVisibility);
-
-          if (keyName != null) {
-            logger.finer('keyName:$keyName');
-            valueJson['encryptionKeyName'] = verbParams[encryptionKeyName];
-            atData.data = jsonEncode(valueJson);
-            // for now skip commit log for keys created through this verb.
-            // If we update/create other keys in the future e.g publickey@alice, @bob:shared_key@alice, change the impl
-            result = await keyStore.put(keyName, atData, skipCommit: true);
-          }
-          response.data = result.toString();
-        } catch (e, st) {
-          logger.warning('$e\n$st');
-        }
+        await _handlePutOperation(
+            verbParams, atSign, keyVisibility, valueJson, response);
         break;
       case 'get':
-        final keyVisibility = verbParams[visibility];
-        final keyNameFromParams = verbParams[keyName];
-        logger.finer('keyVisibility: $keyVisibility');
-        logger.finer('keyNameFromParams: $keyNameFromParams');
-        List<dynamic> result;
-        final filteredKeys = [];
-        if (keyNameFromParams != null && keyNameFromParams.isNotEmpty) {
-          try {
-            final value = await keyStore.get(keyNameFromParams);
-            response.data = value.data;
-            break;
-          } on KeyNotFoundException {
-            throw KeyNotFoundException(
-                'key $keyNameFromParams not found in keystore');
-          }
-        }
-        logger.finer('hasManageAccess: $hasManageAccess');
-        if (keyVisibility != null && keyVisibility.isNotEmpty) {
-          if (hasManageAccess) {
-            result = keyStore.getKeys(
-                regex:
-                    '.*$keyVisibility.*__global$atSign\$|.*$keyVisibility.*__manage$atSign\$');
-          } else {
-            result = keyStore.getKeys(
-                regex: '.*__${keyVisibility}_keys.__global$atSign\$');
-          }
-          for (String key in result) {
-            await _addKeyIfEnrollmentIdMatches(
-                filteredKeys, key, enrollIdFromMetadata!);
-          }
-          var keyMap = {
-            'private':
-                '$enrollIdFromMetadata.$defaultEncryptionPrivateKey.$enrollManageNamespace$atSign',
-            'self':
-                '$enrollIdFromMetadata.$defaultSelfEncryptionKey.$enrollManageNamespace$atSign',
-          };
-
-          var keyString = keyMap[keyVisibility];
-          if (keyString != null) {
-            dynamic value;
-            try {
-              value = await keyStore.get(keyString);
-            } on KeyNotFoundException {
-              logger.warning('key $keyString not found');
-            }
-            if (value != null && value.data != null) {
-              filteredKeys.add(keyString);
-            }
-          }
-
-          response.data = jsonEncode(filteredKeys);
-        }
+        await _handleGetOperation(verbParams, keyVisibility, hasManageAccess,
+            response, enrollIdFromMetadata);
         break;
       case 'delete':
-        final keyNameFromParams = verbParams[keyName];
-        logger.finer('keyNameFromParams: $keyNameFromParams');
-        response.data =
-            (await keyStore.remove(keyNameFromParams, skipCommit: true))
-                .toString();
+        await _handleDeleteOperation(verbParams, response);
         break;
     }
   }
 
-  // Function to add a key to filteredKeys if enrollmentId matches
+  Future<AtData?> _getEnrollData(String key, String atSign) async {
+    try {
+      return await keyStore.get('$key$atSign');
+    } on KeyNotFoundException {
+      logger.warning('enrollment key not found in keystore $key');
+      throw AtEnrollmentException('Enrollment Id $key not found in keystore');
+    }
+  }
+
+  Future<void> _handlePutOperation(
+    HashMap<String, String?> verbParams,
+    String atSign,
+    String? keyVisibility,
+    Map<String, dynamic> valueJson,
+    Response response,
+  ) async {
+    final keyName = _getKeyName(verbParams, atSign, keyVisibility);
+    if (keyName != null) {
+      valueJson['encryptionKeyName'] = verbParams[encryptionKeyName];
+      final atData = AtData()..data = jsonEncode(valueJson);
+      final result = await keyStore.put(keyName, atData, skipCommit: true);
+      response.data = result.toString();
+    }
+  }
+
+  Future<void> _handleGetOperation(
+    HashMap<String, String?> verbParams,
+    String? keyVisibility,
+    bool hasManageAccess,
+    Response response,
+    String enrollIdFromMetadata,
+  ) async {
+    final keyNameFromParams = verbParams[keyName];
+    var atSign = AtSecondaryServerImpl.getInstance().currentAtSign;
+    if (keyNameFromParams != null && keyNameFromParams.isNotEmpty) {
+      try {
+        final value = await keyStore.get(keyNameFromParams);
+        response.data = value.data;
+        return;
+      } on KeyNotFoundException {
+        throw KeyNotFoundException(
+            'key $keyNameFromParams not found in keystore');
+      }
+    }
+    final filteredKeys = await _getFilteredKeys(
+        keyVisibility, hasManageAccess, enrollIdFromMetadata, atSign);
+    response.data = jsonEncode(filteredKeys);
+  }
+
+  /// If current enrollment has __manage access then return both __global and __manage keys with visibility [keyVisibility]
+  /// Otherwise return only __global keys with visibility [keyVisibility]
+  /// Also return the encrypted default encrption private key and encrypted self encryption key for enrollmentId [enrollIdFromMetadata]
+  Future<List<String>> _getFilteredKeys(String? keyVisibility,
+      bool hasManageAccess, String enrollIdFromMetadata, String atSign) async {
+    final result = keyVisibility != null && keyVisibility.isNotEmpty
+        ? hasManageAccess
+            ? keyStore.getKeys(
+                regex:
+                    '.*$keyVisibility.*__global$atSign\$|.*$keyVisibility.*__manage$atSign\$')
+            : keyStore.getKeys(
+                regex: '.*__${keyVisibility}_keys.__global$atSign\$')
+        : <String>[];
+
+    final filteredKeys = <String>[];
+    for (final key in result) {
+      await _addKeyIfEnrollmentIdMatches(
+          filteredKeys, key, enrollIdFromMetadata);
+    }
+
+    final keyMap = {
+      'private':
+          '$enrollIdFromMetadata.$defaultEncryptionPrivateKey.$enrollManageNamespace$atSign',
+      'self':
+          '$enrollIdFromMetadata.$defaultSelfEncryptionKey.$enrollManageNamespace$atSign',
+    };
+
+    final keyString = keyMap[keyVisibility];
+    if (keyString != null) {
+      try {
+        final value = await keyStore.get(keyString);
+        if (value?.data != null) {
+          filteredKeys.add(keyString);
+        }
+      } on KeyNotFoundException {
+        logger.warning('key $keyString not found');
+      }
+    }
+    return filteredKeys;
+  }
+
+  Future<void> _handleDeleteOperation(
+    HashMap<String, String?> verbParams,
+    Response response,
+  ) async {
+    final keyNameFromParams = verbParams[keyName];
+    response.data =
+        (await keyStore.remove(keyNameFromParams, skipCommit: true)).toString();
+  }
+
+  /// List only keys from current enrollment. Do not list keys from another enrollment.
+  /// Get the valueJson from keystore for [key].
+  /// If the enrollment in valueJson matches [enrollIdFromMetadata], then add [key] to [filteredKeys]
   Future<void> _addKeyIfEnrollmentIdMatches(List<dynamic> filteredKeys,
       String key, String enrollIdFromMetadata) async {
     final value = await keyStore.get(key);
@@ -168,6 +192,11 @@ class KeysVerbHandler extends AbstractVerbHandler {
     }
   }
 
+  /// Key structure varies based on visibility. Construct and return the key name based on [keyVisibility]
+  /// Key name for public visibility - 'public:<keyname>.__public_keys.<namespace>@<atsign>'
+  /// Key name for private visibility - 'private:<appName>.<deviceName>.<keyname>.__private_keys.<namespace>@<atsign>'
+  /// Key name for self visibility  - '<appName>.<deviceName>.<keyname>.__self_keys.<namespace>@<atsign>'
+  /// returns null, if [keyVisibility] is not public|private|self
   String? _getKeyName(HashMap<String, String?> verbParams, String atSign,
       String? keyVisibility) {
     if (keyVisibility == 'public') {
